requirements.txt
aiofiles==22.1.0
aiosqlite==0.17.0
certifi==2022.9.24
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
httptools==0.5.0
iso8601==1.1.0
multidict==6.0.2
pypika-tortoise==0.1.6
python-dateutil==2.8.2
pytz==2022.6
sanic==22.9.1
sanic-routing==22.8.0
six==1.16.0
tortoise-orm==0.19.2
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvloop==0.17.0
websockets==10.4
asyncpg
